Task2/Egyptian Hieroglyphs/requirements.txt
streamlit==1.22.0
tensorflow==2.13.0
numpy==1.24.2
Pillow==9.4.0
scikit-learn==1.2.2
joblib==1.2.0
opencv-python-headless==4.7.0.72